HasControlPanelUI
