+++ src/Element/Concerns/Revisionable.php
@@ -13,9 +13,18 @@ use Illuminate\Support\Facades\DB;
 
 use function CraftCms\Cms\t;
 
-/** @phpstan-ignore trait.unused */
+/**
+ * @internal
+ */
 trait Revisionable
 {
+    /**
+     * @var int|null The ID of the revision’s row in the `revisions` table
+     *
+     * @since 3.2.0
+     */
+    public ?int $revisionId = null;
+
     /**
      * @var int|null The creator’s ID
      */
